Reporter selection logic excluded from main server file
